+++ packages/core/src/transports/mempool/getTransactions.ts
@@ -1,0 +1,76 @@
+import type { RpcMethodHandler } from '../types.js'
+import type {
+  MempoolBalanceResponse,
+  MempoolUTXOTransactionsResponse,
+} from './mempool.types.js'
+
+export const getTransactions: RpcMethodHandler<'getTransactions'> = async (
+  client,
+  { baseUrl },
+  { address, limit = 50, offset = 0 }
+) => {
+  async function* generator() {
+    const apiUrlAddress = `${baseUrl}/address/${address}`
+    const balanceResponse = (await client.request({
+      url: apiUrlAddress,
+      fetchOptions: { method: 'GET' },
+    })) as unknown as MempoolBalanceResponse
+    const totalTxns =
+      balanceResponse.chain_stats.tx_count +
+      balanceResponse.mempool_stats.tx_count
+    let currentOffset = offset
+    let after_txid: string | undefined = undefined
+    while (currentOffset < totalTxns) {
+      const page = Math.floor(currentOffset / limit) + 1
+      const apiUrl = `${baseUrl}/address/${address}/txs${after_txid ? `?after_txid=${after_txid}` : ''}`
+
+      const response = (await client.request({
+        url: apiUrl,
+        fetchOptions: { method: 'GET' },
+      })) as unknown as MempoolUTXOTransactionsResponse
+
+      const transactions = response.map((utxo) => ({
+        hash: utxo.txid,
+        txid: utxo.txid,
+        vout: utxo.vout.map((vout, index) => ({
+          n: index,
+          scriptPubKey: {
+            address: vout.scriptpubkey_address,
+            asm: vout.scriptpubkey_asm,
+            type: vout.scriptpubkey_type,
+            desc: vout.scriptpubkey,
+            hex: vout.scriptpubkey,
+          },
+          value: vout.value,
+        })),
+        vin: utxo.vin.map((vin) => ({
+          scriptSig: {
+            asm: vin.scriptsig_asm,
+            hex: vin.scriptsig,
+          },
+          sequence: vin.sequence,
+          txinwitness: vin.witness,
+          txid: vin.txid,
+          vout: vin.vout,
+        })),
+      }))
+
+      yield {
+        transactions,
+        total: totalTxns,
+        page,
+        itemsPerPage: limit,
+      }
+
+      currentOffset += response.length
+      after_txid = response[response.length - 1]?.txid
+      if (response.length < limit) {
+        break
+      }
+    }
+  }
+
+  return {
+    result: generator(),
+  }
+}
